improved train_k
